Rework ApiStub
