--- 1.19.4/src/main/java/opekope2/lilac/stub/ApiStub.java
@@ -1,16 +1,0 @@
-package opekope2.lilac.stub;
-
-import opekope2.lilac.api.ILilacApi;
-import opekope2.lilac.impl.LilacApi;
-import org.jetbrains.annotations.NotNull;
-
-public final class ApiStub {
-    private ApiStub() {
-    }
-
-    @SuppressWarnings("unused")
-    @NotNull
-    public static ILilacApi getApiImplementation() {
-        return LilacApi.INSTANCE;
-    }
-}
