+++ 1.19.4/src/main/kotlin/opekope2/lilac/api/ILilacApi$Instance.kt
@@ -1,0 +1,9 @@
+package opekope2.lilac.api
+
+import opekope2.lilac.impl.LilacApi
+
+@Suppress("unused", "ClassName")
+private object `ILilacApi$Instance` {
+    @JvmStatic
+    fun get(): ILilacApi = LilacApi
+}
